+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## [0.20.0] - 2024-09-25
 
 ### Added
 
@@ -178,6 +178,7 @@
 
 - Initial release
 
+[0.20.0]: https://github.com/tinted-theming/tinty/compare/v0.19.0...v0.20.0
 [0.19.0]: https://github.com/tinted-theming/tinty/compare/v0.18.0...v0.19.0
 [0.18.0]: https://github.com/tinted-theming/tinty/compare/v0.17.0...v0.18.0
 [0.17.0]: https://github.com/tinted-theming/tinty/compare/v0.16.0...v0.17.0